+++ src/components/DownloadPane.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+  <div>下载页面</div>
+</template>
