+++ src/components/WatermarkComponents/RemoveProgress.vue
